requirements.txt
openai==1.54
python-telegram-bot==21.7